% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicMC.R
\name{absorbing}
\alias{absorbing}
\title{Function to identify the absorbing states}
\usage{
absorbing(trMatrix)
}
\arguments{
\item{trMatrix}{Transition matrix of the Markov chain}
}
\value{
The absorbing states of the finite-state discrete time-homogeneous Markov Chain
}
\description{
This method is used to identify the absorbing states of a discrete time-homogeneous Markov chain.
}
\examples{
egstate <- c("a", "b", "c")
MC <- matrix(data = c(1, 0, 0, 0, 1, 0, 0, 0, 1),
         nrow = 3, byrow = TRUE,
         dimnames = list(egstate, egstate))

absorbing(MC)

}
\references{
Grimmett, G., & Stirzaker, D. (2001). Probability and Random Processes (3rd ed.). New York: Oxford University Press.
}
