% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicMC.R
\name{condDist}
\alias{condDist}
\title{Obtain the conditional distribution of the states}
\usage{
condDist(trMatrix, initialstate = NULL, steps)
}
\arguments{
\item{trMatrix}{Transition matrix of the Markov chain}

\item{initialstate}{A vector of distribution of the initial states.}

\item{steps}{The number of steps}
}
\value{
A vector of n-step transition probabilities of the states given a distribution of initial states
}
\description{
Function to calculate the conditional distribution of states given a distribution of initial states.
}
\examples{
egstate <- c("a", "b", "c")
MC <- matrix(data = c(1, 0, 0, 0, 1, 0, 0, 0, 1),
         nrow = 3, byrow = TRUE,
         dimnames = list(egstate, egstate))

initial <- c(0.5,0.5,0)
n <- 5
condDist(MC,initial,n)

}
\references{
Grimmett, G., & Stirzaker, D. (2001). Probability and Random Processes (3rd ed.). New York: Oxford University Press.
}
