% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicMC.R
\name{displaytrM}
\alias{displaytrM}
\title{Display the transition matrix}
\usage{
displaytrM(trMatrix)
}
\arguments{
\item{trMatrix}{Transition matrix of the Markov chain}
}
\value{
The transition matrix of the corresponding Markov chain
}
\description{
This method is used to display the transition matrix of a finite-state discrete time-homogeneous Markov Chain.
}
\examples{
egstate <- c("a", "b", "c")
MC <- matrix(data = c(1, 0, 0, 0, 1, 0, 0, 0, 1),
         nrow = 3, byrow = TRUE,
         dimnames = list(egstate, egstate))

displaytrM(MC)


}
\references{
Chambers, J. (2008). Software for Data Analysis: Programming with R. New York: Springer Science+Business Media, LLC.
}
