% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ftse.R
\docType{data}
\name{ftse}
\alias{ftse}
\title{Data from ftse}
\format{A data frame with 32 rows and 3 variables:
\describe{
  \item{Serial.number}{Serial number}
  \item{Date}{Date of the closing price}
  \item{Closing.price}{Closing price, from 1..32}
}}
\source{
https://www.londonstockexchange.com/exchange/prices-and-markets/stocks/indices/summary/summary-indices.html?index=UKX
}
\usage{
data(ftse)
}
\description{
This is data of the closing price of FTSE 100 from 1 October, 2018 to
13 November, 2018
}
\examples{
data(ftse)
}
\keyword{datasets}
