% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicMC.R
\name{simMC}
\alias{simMC}
\title{Simulation of a finite-state discrete time-homogeneous Markov Chain}
\usage{
simMC(trMatrix, InitStateIndex = NULL, steps, presentation = TRUE)
}
\arguments{
\item{trMatrix}{Transition matrix of the Markov chain}

\item{InitStateIndex}{The index of the initial state in the vector of states}

\item{steps}{The number of iterations}

\item{presentation}{Decide whether the states simulated is in number or state name.The default value is TRUE.}
}
\value{
A sequence of states
}
\description{
This method is used to simulate a finite-state discrete time-homogeneous Markov Chain.
}
\examples{
egstate <- c("1", "2", "3")
MC <- matrix(data = c(1, 0, 0, 0, 1, 0, 0, 0, 1),
         nrow = 3, byrow = TRUE,
         dimnames = list(egstate, egstate))

index <- 1
n <- 50
simMC(MC,index,n)

}
\references{
Grimmett, G., & Stirzaker, D. (2001). Probability and Random Processes (3rd ed.). New York: Oxford University Press.
}
