% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicMC.R
\name{stationary}
\alias{stationary}
\title{Function to obtain the stationary distribution of a Markov Chain}
\usage{
stationary(trMatrix)
}
\arguments{
\item{trMatrix}{Transition matrix of the Markov chain}
}
\value{
A vector of stationary distribution corresponding to the Markov chain.
}
\description{
This method is used to get the stationary distribution of a finite discrete time-homogeneous Markov chain.
}
\examples{
egstate <- c("a", "b", "c")
MC <- matrix(data = c(1, 0, 0, 0, 1, 0, 0, 0, 1),
         nrow = 3, byrow = TRUE,
         dimnames = list(egstate, egstate))

stationary(MC)

}
\references{
Grimmett, G., & Stirzaker, D. (2001). Probability and Random Processes (3rd ed.). New York: Oxford University Press.
}
