% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicMC.R
\name{stock}
\alias{stock}
\title{Convert the data into a transition matrix for stock market}
\usage{
stock(y, pricecolumn, interval = 10)
}
\arguments{
\item{y}{Data of the stock prices}

\item{pricecolumn}{The index of the column containg the prices}

\item{interval}{The range for the state "non-significant". The default value is 10. If interval = 0, "non-significant" state will exist when no price change.}
}
\value{
The transition matrix of the Markov chain
}
\description{
This method is used to create a transition matrix from raw data of stock prices.
}
\examples{
data(ftse)
stock(ftse,3,5)
}
