% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicMC.R
\name{transient}
\alias{transient}
\title{Function to identify transient states}
\usage{
transient(trMatrix)
}
\arguments{
\item{trMatrix}{Transition matrix of the Markov chain}
}
\value{
The transient states of a finite-state discrete time-homogeneous Markov Chain.
}
\description{
This method is used to identify the transient states of a finite-state discrete time-homogeneous Markov Chain.
}
\examples{
egstate <- c("a", "b", "c")
MC <- matrix(data = c(1, 0, 0, 0, 1, 0, 0, 0, 1),
         nrow = 3, byrow = TRUE,
         dimnames = list(egstate, egstate))

transient(MC)

}
\references{
Grimmett, G., & Stirzaker, D. (2001). Probability and Random Processes (3rd ed.). New York: Oxford University Press.
}
