% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicMC.R
\name{trProb}
\alias{trProb}
\title{Calculate the n-step transition matrix}
\usage{
trProb(trMatrix, steps)
}
\arguments{
\item{trMatrix}{Transition matrix of the Markov chain}

\item{steps}{The number of steps}
}
\value{
A matrix of n-step transition probabilities of the states
}
\description{
Function to calculate the n-step transition matrix of a finite-state discrete time-homogeneous Markov Chain.
}
\examples{
egstate <- c("a", "b", "c")
MC <- matrix(data = c(1, 0, 0, 0, 1, 0, 0, 0, 1),
         nrow = 3, byrow = TRUE,
         dimnames = list(egstate, egstate))

n <- 5
trProb(MC,n)

}
\references{
Grimmett, G., & Stirzaker, D. (2001). Probability and Random Processes (3rd ed.). New York: Oxford University Press.
}
