% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsoton.R
\docType{data}
\name{wsoton}
\alias{wsoton}
\title{Data from wsoton}
\format{A data frame with 35 rows and 2 variables:
\describe{
  \item{Dates}{Date of the data collected}
  \item{Weather}{Weather type, from 1..35}
}}
\source{
https://www.worldweatheronline.com/southampton-weather-history/hampshire/gb.aspx
}
\usage{
data(wsoton)
}
\description{
This is data of the weather of Southampton, UK from 1 October, 2018 to
4 November, 2018
}
\examples{
data(wsoton)
}
\keyword{datasets}
